+++ nengo_loihi/simulator.py
@@ -278,7 +278,7 @@ class Simulator:
         """Copy all probed signals to buffers."""
         self._probe_step_time()
 
-        for probe in self.model.probes:
+        for probe in self.model.nengo_probes:
             if probe in self.model.chip2host_params:
                 continue
             assert probe.sample_every is None, "probe.sample_every not implemented"
@@ -454,7 +454,7 @@ class StepRunner:
             error_target = receiver.error_target
             assert error_target is not None
 
-            conn = self.model.probe_conns[error_target]
+            conn = self.model.nengo_probe_conns[error_target]
             error_synapse = self.model.objs[conn]["decoders"]
             assert error_synapse.learning